feat : 수익률 계산 처리

+++ src/main/java/lotto/domain/purchase/lottocount/LottoCountCalculator.java
@@ -3,7 +3,7 @@ package lotto.domain.purchase.lottocount;
 import static lotto.domain.purchase.lottocount.CashPolicy.PRICE_PER_LOTTO;
 
 public final class LottoCountCalculator {
-    public static Integer divideByLottoPrice(Integer cash) {
-        return cash / PRICE_PER_LOTTO;
+    public static Integer divideByLottoPrice(Long cash) {
+        return Math.toIntExact(cash / PRICE_PER_LOTTO);
     }
 }
